requirements.txt
lightning
torch
pyyaml
pydantic
tokenizers
transformers
accelerate